Move output functions to new writers module
