+++ src/planingfsi/fsi/simulation.py
@@ -11,7 +11,7 @@ from .. import config
 from .. import logger
 from ..dictionary import load_dict_from_file
 from ..fe.structure import StructuralSolver
-from ..general import write_as_dict
+from .. import writers
 from ..potentialflow.solver import PotentialPlaningSolver
 from .interpolator import Interpolator
 
@@ -254,7 +254,7 @@ class Simulation:
 
     def write_results(self) -> None:
         if self.check_output_interval() and not config.io.results_from_file:
-            write_as_dict(
+            writers.write_as_dict(
                 os.path.join(self.it_dir, "overallQuantities.txt"),
                 ["Ramp", config.ramp],
                 ["Residual", self.solid_solver.res],
