+++ src/planingfsi/potentialflow/solver.py
@@ -13,6 +13,7 @@ from .. import config
 from .. import general
 from .. import logger
 from .. import solver
+from .. import writers
 from ..fsi import figure
 from ..fsi import simulation as fsi_simulation
 from .pressureelement import PressureElement
@@ -360,7 +361,7 @@ class PotentialPlaningSolver:
 
     def _write_forces(self) -> None:
         """Write forces to file."""
-        general.write_as_dict(
+        writers.write_as_dict(
             self.simulation.it_dir / f"forces_total.{config.io.data_format}",
             ["Drag", self.drag_total],
             ["WaveDrag", self.drag_wave],
@@ -377,7 +378,7 @@ class PotentialPlaningSolver:
 
     def _write_pressure_and_shear(self) -> None:
         """Write pressure and shear stress profiles to data file."""
-        general.write_as_list(
+        writers.write_as_list(
             self.simulation.it_dir / f"pressureAndShear.{config.io.data_format}",
             ["x [m]", self.x_coord],
             ["p [Pa]", self.pressure],
@@ -386,7 +387,7 @@ class PotentialPlaningSolver:
 
     def _write_free_surface(self) -> None:
         """Write free-surface profile to file."""
-        general.write_as_list(
+        writers.write_as_list(
             self.simulation.it_dir / f"freeSurface.{config.io.data_format}",
             ["x [m]", self.x_coord_fs],
             ["y [m]", self.z_coord_fs],
